Remove unneeded CMake settings (assume at least open62541 v1.3)

+++ backends/stdout/examples/backend.h
@@ -7,10 +7,12 @@
 
 #ifndef __BACKEND_H__
 #define __BACKEND_H__
+
 #include <NodesetLoader/NodesetLoader.h>
 #include <stdbool.h>
 #include <stdio.h>
 
-int addNamespace(void *userContext, const char *namespaceUri);
+unsigned short addNamespace(void *userContext, const char *namespaceUri);
 void dumpNode(void *userContext, const NL_Node *node);
+
 #endif
